app/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.2
dnspython==2.5.0
ecdsa==0.18.0
email-validator==2.1.0.post1
PyJWT==2.10.1
exceptiongroup==1.2.0
fastapi==0.109.2
h11==0.14.0
idna==3.6
passlib==1.7.4
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pymongo==4.6.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.1
